Redesigned artifact data classes.
Moved the artifact data of each derived item out to a text file. A
simple parser will be written that will read that information into the
application. This is to make artifacts not be hard-coded. Other
hard-coded data, like locations, etc., will also be moved out to text
files eventually.