Redesigned Location and Province classes.
Moved all of the location data and some of the province data out to a
text file. A parser will be written to read it back in at runtime so
that the locations are not hard-coded. It is not clear yet how these
changes will affect quest coding or map icon coding once those two tasks
are undertaken.