Added character class parser functionality.
The character class parser can now read in the "classes.txt" file and
produce a vector of CharacterClass objects. This vector would be most
useful in the class selection screen during character creation. It
should display a list of classes, and each class should have a tooltip
when hovered over with the mouse.