Modified texture manager. Added character panel.
A blank draft of the character sheet was added; pressing Tab in the game
world goes to the sheet, and pressing Escape in the sheet goes back to
the game world. The texture manager now takes filenames directly, and is
intended to be used with the TextureFile class for hard-coded textures
now. This made the texture manager class more lightweight. Preloading
all cinematics is now an option in the texture manager. Added
"Debug::mention()" for optional debug messages. Replaced
"Renderer::checkSuccess()" with a call to "Debug::check()". Removed wide
lens images from the "samples" folder (I don't think the program would
need to use the wide lens camera model for now).