Started OpenCL program integration.
Added the CLProgram class for the OpenCL program instance to be stored
in the GameData object. The game world background is now a rainbow
gradient instead of all black. The gradient is a debug surface
indicating that communication between the host and the graphics card is
working. Pressing left click or E in the game world now plays some
placeholder sounds like weapon clank/swish and open door.