Added camera code and some rendering code.
The background colors are now temporarily determined by the player's
camera and the ray directions. This is for debug purposes to verify the
correctness of the camera rotation code (and it has shown to be
correct). In the game world panel, holding down the left or right mouse
button and moving the mouse will make the background colors rotate,
showing the camera rotation (also for debug purposes). Added Options
class, which holds several values like screen resolution and mouse
sensitivity. It will eventually load from an options text file.