Added Matrix4 method to CoordinateFrame class.
The CoordinateFrame class can now produce a Matrix4 representation of
its 3D axis with a given point acting as the frame's origin. Also added
the original Arena's screen width and height dimensions to the math
constants file. Also removed the Algorithms folder; it wasn't doing very
much. It might be added back at a later time. Also reorganized my to-do
list so I have a better idea of what to work on.