Added character portrait selection. Added skip intro option.
The ChooseAttributesPanel now allows the player to select what portrait
to use for their chosen race, class, and gender combination. Also added
vertical field of view functionality, so the value in the options file
works now. Also added "SkipIntro" value in the options file, which
causes the program to jump to the main menu on startup if true. Also
added compass slider to the game world interface. It doesn't move with
the player's camera yet, but it shouldn't be too hard to figure out.