Start support for loading Arena's IMG files
Currently supports only uncompressed and "type-4" compression. IMG files that
use "type-8" compression, or which are actually "headerless"/raw, are currently
unsupported.

Also note that due to the original game using 8-bit color, the correct palette
needs to be used for conversion. Some IMG files contain their own palette, but
otherwise the correct palette needs to be selected when loading it (which
currently only the default PAL.COL palette is used; the others are DREARY.COL,
DAYTIME.COL, and CHARSHT.COL).