Use TAMRIEL.MNU and STARTGAM.MNU
There is one extra MNU file, MAPBTNS.MNU, which is a bit interesting. As far as
I can tell, it's actually a selection lookup table for the race select screen.
The mouse cursor's X/Y position in the 320x200 screen is used to get the 8-bit
pixel value from MAPBTNS.MNU, and the value is an index of sorts that indicates
the province/race to select.