Added attributes texture to ChooseAttributesPanel.
The ChooseAttributesPanel now uses the character stats background
instead of a solid color, though the palette isn't right yet (it should
be CHARSHT.COL). The "done" button works in it now, too. The
CharacterPanel should use a different texture that has the "Next Page"
button, but it hasn't been found yet. The parameters in the
ChooseAttributesPanel constructor were reordered to follow the order of
character creation.