Organized texture manager palette usage.
All palettes are now stored in the texture manager. The palettes are
paired with a PaletteName for easier filename mapping and convenience.
Whenever loading a surface or texture, there is an optional PaletteName
parameter for loading with a certain palette (default is PAL.COL). This
change does not affect images with a built-in palette.