Added textures to test world.
Updated data folder. The test world now has some textures and... arches.
The textures are hard-coded at the moment, but this is just for tests
anyway, and to show that transparency is working in the ray tracer (not
partial transparency, though). Also fixed incorrect bit shifts in
Float3::fromRGB() and Float4::fromARGB().