Fix the unordered_map for Surfaces
Being able to use std::hash on an enum is a C++14 feature. And since the point
of a hash is to generate a unique integer for an object of a given type, there
doesn't seem to be a real reason to pass an integer through it (the integer
will be unique already).

Also it seems GCC/Clang don't like declaring an unordered_map using an
incomplete type.