Changed sound filenames to use the originals.
The SoundName enumerations now point to the .VOC files in the ARENA
folder and GLOBAL.BSA instead of the tool-extracted ones I planned on
using. I also removed the speech filenames since those are out of scope
right now (just focusing on the floppy disk data for the time being). I
also removed "SoundName::Dungeon5" because it's a duplicate of
"Dungeon1".