Added music and sound buttons in pause menu.
The volumes for music and sound can be changed in the pause menu now. No
sounds play yet, so only changing the music volume will have an audible
effect for now. The changes do not save to options.txt yet, so they are
lost on program exit. Editing options.txt itself is an option, though.