Added class skeletons for loading Arena assets.
The texture manager will no longer have all of the image reading and
decompression code since it will be moved to various classes designed
for decompressing their given format. The interface is still the same
with the rest of the program, but the texture manager will now interact
with the asset classes behind the scenes.