Modified texture manager usage of PaletteNames.
The texture manager now takes a std::string for the palette to use,
instead of a PaletteName enum. This allows for using the palettes
contained in IMG files. The file extension is taken from the string, and
that determines how the palette is obtained. If the extension is ".COL",
the file is read in and used. If it is ".IMG" or ".MNU", the file is
loaded for its built-in palette, which is then extracted using
IMGFile::extractPalette(). Also added new PaletteFile class for which to
obtain the palette string from, given a PaletteName.