Removed VoxelMaterialType.
The VoxelMaterialType was adding too much unnecessary complexity to the
rough draft voxel design. All the water and lava in the game is always
at the y=0 level, that is, the lowest plane of blocks in the map, so
there doesn't need to be a specific material for every voxel everywhere.
Also added a link in FLCFile.h for a website describing the FLIC format.
Also moved the VoxelReference class to the Rendering folder since it's
probably going to be used with OpenCL.