Revised class creation interface.
The interface for class creation now more closely resembles the
original. However, the palette used for the pop-up texture doesn't seem
to be correct since the palette extraction code is apparently giving the
wrong colors. Also changed the width of spaces for FontName::C. Also
made a slight change to the class tooltip in ChooseClassPanel.