Added Sprite class and several design comments.
I'm currently theorizing about the sprite design I want to use. It
involves managing which voxels in the world are touched by which
sprites, and efficiently updating the geometry in kernel memory to
reflect those changes each frame. Read my long paragraphs of comments in
CLProgram.h and SpriteReference.h for more details.