Changed some types in Image and Cinematic panels.
The PaletteName member in the ImagePanel and CinematicPanel was changed
to a std::string to have the same flexibility in palette usage as
regular textures have. This change was made to the TextureName member in
the ImagePanel as well, but not the TextureSequenceName member in the
CinematicPanel yet because FLC file loading and the
TextureManager::getTextures() method aren't implemented yet.