Added texture loading support for RCI files.
The two .RCI files, WATERANI.RCI and LAVAANI.RCI, are screen-space
animations for water and lava, respectively. I don't think they will be
used, though, unless the player camera is restricted to only horizontal
movement like in the original game, in which case the RCI images would
be stretched to fill the bottom half of the frame buffer.