Added ImageSequencePanel class, edited palette extraction.
The ImageSequencePanel class shares similar functionality with the
CinematicPanel, but it is designed for displaying a set of still images
instead of a quickly animating video. It allows the user to click or
press enter/space to skip an image, or escape to skip the whole
sequence. Also removed unnecessary texture sequence names. Also cleaned
up the IMGFile::extractPalette() code some, but it's still not correct.
Perhaps double palette indirection is involved for some images?