Removed dependency on tool-extracted cinematics.
Updated data folder. The program is now using more of the original Arena
game files by loading cinematics from FLC files instead of
tool-extracted images in the textures folder. However, the frame
decoding functionality is not complete yet, so cinematics will be black
for now. Also modified the CinematicPanel and TextCinematicPanel
constructors so they take a filename instead of a texture sequence enum.