Trimmed unnecessary debug code from SETFile.
There doesn't need to be an explicit debug message about a chunk index
being out of range because the user should call SETFile::getImageCount()
beforehand to ensure all requests are within bounds. Also modified the
Readme a little bit, and made the scope more general because the
Projects tab is being used now.