More work on FLC file decoding.
Trimmed a bit of the excess code and wrote chunk decoding methods for
better organization. Now instead of cinematics being black, they're an
animating block of colored noise. Just need to figure out what exactly
the RLE encoding method for the frames is. Each image is supposedly
compressed on a per-line basis.