Added type 4 and 8 CIF support, added character faces.
It turns out that the existing Compression::decodeType04() and
decodeType08() methods work with the associated CIF files, so now all
CIF files can be decoded. Also, there are two versions of every
character face (for some reason); one is the "full" image for the
character sheet, and one is the "trimmed" image used in the box in the
game world interface.