Fixed crash when returning to main menu.
The player data for a game session gets deleted when returning to the
main menu, but the pause menu still renders one more frame before then
and it needs the session data to find which head portrait to use in the
corner. This change saves the filename and portrait index beforehand so
the game data doesn't need to be referenced later. I still think this
design could be improved.