Switched to rough draft horizontal-only camera.
I commented out the 3D camera code and replaced it with a rough draft
that mostly matches the turning behavior of Arena. It's still terrible
like the original! I do not like this awkward turning acceleration at
all; it's identical to playing a first-person RPG with a joystick! As
stated in a comment in the code, I can't wait to add back an option for
a more Daggerfall-like camera in the future.