Added support for FLC file decoding.
It was quite a struggle to decode this file format. The conflicting
documentation between FLIC websites didn't help either. I had to look
through a hex print of some FLC files to find that chunk headers are 6
bytes, not 16! At least the movies like the opening scroll and Ria
Silmane work now, so the headache was worth it.