Added more game interface behavior.
The various buttons on the in-game interface can now be clicked, and a
couple of them already go to their respective menu (others just print a
console message). The left and right invisible rectangles in the screen
for camera rotation also have proper behavior now (the other regions do
nothing yet).