Properly clear original frame buffer.
With every texture using its alpha component automatically now, this
causes images to remain in the background after switching panels if the
current image contains alpha (which most of them do). This change makes
sure that no remnants of past frames persist between panels. Also, the
Travis CI build is failing because SDL 2.0.5 isn't in the distribution
yet, so I guess I just have to wait for that, maybe.