Update .travis.yml
Package repos will be lagging behind on SDL 2.0.5 for a while (as we now require some functionality specific to 2.0.5 in TESArena now), so we clone via mercurial the 2.0.5 release and build it.  It should work, but we won't fully know until we merge and have Travis test within its environment.