Added rough draft screenshot capability for game world.
Pressing the print screen key in the game world will save a screenshot
to the executable folder with the name "out.bmp". There might be a
redesign of the event handling coming up so that the screenshot code
would only need to be in one place and not in every panel. Also removed
the Renderer::getFormattedARGB() method since the window format isn't
important anymore when manipulating SDL_Surfaces (since they shouldn't
interact).