Fixed some edge cases in IMGFile.
A couple IMG files like HEDGE.IMG were not readable because their first
few pixels are black, so the header would just be zeroes and the
dimensions would be incorrect. This change assumes that all IMG files
that are 4096 bytes are 64x64 wall textures. I don't believe this will
cause any regression.