Moved CLProgram from GameData to Renderer.
The OpenCL object is now owned by the Renderer class. It originally made
sense to only have the world rendering initialized when the game data
object is alive, but having it in the renderer instead means that more
similar functionality will be closer together. To me, the Renderer class
was originally for wrapping SDL renderer operations, but now it also
wraps 3D rendering. The CLProgram::render() method now also returns a
pointer to the game world frame buffer instead of updating a SDL texture
itself.