Added wrappers for SDL primitive drawing functions.
The new renderer methods wrap SDL drawing functions for colored pixels,
lines, and rectangles. One thing I forgot to mention in the previous
commit is that the game world view is now centered above the interface
as it was in the original game, instead of being centered in the middle
of the window.