Added Texture wrapper for use with texture manager.
This change lets all SDL_Textures stored in the texture manager be used
by the application as references, and the wrapper also adds new methods
for getting the width and height of the owned texture. This is mostly a
convenience feature so the SDL_QueryTexture() function doesn't need to
be called. Also, the performance of cursor rendering should be a tiny
bit faster because they're hardware-accelerated now (instead of being
turned into SDL_Textures on the fly).