Refactored ListBox class.
The ListBox no longer inherits from Surface, and now uses an SDL_Texture
for displaying its text. Updating the texture is now more efficient
because each of the text boxes are stored for the lifetime of the list
box, so allocations are no longer done on the fly when scrolling through
the list box. Also removed some unnecessary forward declarations in
TextureManager.h.