Trimmed Surface functionality.
I'm continuing to phase out the Surface class since the program is more
oriented towards SDL_Textures now, and SDL_Surfaces are just being used
as scratch images which eventually get converted into hardware textures
for rendering. The Surface class is now more like the Texture class: a
lightweight wrapper. It also should've been in the Rendering folder all
along, so that's where it is now.