Added support for CEL file decoding.
CEL video files share nearly all of their functionality with FLC files,
so this was an easy change to make. The three CEL files used in Arena
are for the short bursts of magic cast into the night sky during class
creation (if you mess with the palette for the night sky texture, it
shows clear outlines for the warrior/mage/rogue constellations; I wonder
how those are drawn over the CEL animations).