Added CLProgram::zeroBuffer().
After some debugging, it appears that certain OpenCL implementations do
not zero out cl::Buffers upon allocation. This means that, on certain
drivers, garbage values for voxel references are read in the kernel
causing severe out-of-bounds errors to occur. This is no longer a
problem because the suspected buffers are explicitly zeroed out now.