Added rough draft sprite code.
This change adds a lot of code for letting the OpenCL object manage
sprites in device memory. There aren't any sprites yet, but I will be
testing this code soon to see if it functions as intended. This current
iteration is somewhat disorganized and does redundant checks in some
places, but the rendering performance should be much better than that of
my test project.