Added sprite rendering and some test sprites.
Updated data folder. There are now a couple bartenders scattered around
the test world. They animate and face towards the player. When hundreds
of sprites occupy the world, the CPU is heavily bound by calls to
clEnqueueWriteBuffer(); this is to be expected since this is a first
attempt at better sprite communication with device memory, and it can
eventually be optimized. The sprite rendering is fast and reference
management appears to be correct, nonetheless.