Improved geometry updating performance.
The CLProgram now queues up voxel and sprite updates during a frame and
pushes them all to device memory at once using non-blocking
(asynchronous) clEnqueueWriteBuffer() calls. This results in the program
being much less CPU bound when updating hundreds of voxels and sprites
simultaneously. The writes can be done in parallel because they all
point to unique regions in device memory. It's not perfect yet because
it's still a little wasteful of work, but it's significantly better when
updating geometry in bulk.