Separated some audio manager functionality.
The filenames for music and sound files are now in their own classes
(MusicFile and SoundFile) so that the audio manager can worry about just
the filenames themselves, instead of also accessing mappings for
MusicNames and SoundNames. Also removed the MusicType enum since it
probably won't end up being as useful as I thought. Also added
MusicName::Dungeon5 since it's not a duplicate of Dungeon1 if the .XFM
version is used.