Added tooltips for game interface buttons.
The buttons in the game world interface now have tooltips shown on the
left side of the screen when hovered over. The original game doesn't
have tooltips, and its icons aren't very clear to begin with, usually
requiring the player to either find them in the manual or just plain
experiment with them.