Renamed upper-level data folder to data_files.
It was a little misleading to have two folders named 'data', so the one
at the top level of the repository has been renamed to 'data_files'. The
'data' and 'options' folders inside it should be brought up one folder
if running the executable at the top-level directory of the repository
(i.e., on Linux).