Added a little more light updating code.
I've gotten a bit lost in my own design here because the light owners
idea was kind of tacked on in the process. I'll need to rethink some of
this and reorganize some code, and maybe get rid of a few of the
seventeen unordered maps I'm using! This is my first attempt at dynamic
geometry management in a uniform grid, so it's not very efficient, but
at least it's correct.