Finished Arena-like mouse look and movement.
The methods for handling mouse and keyboard input in the game world have
been transformed into methods for explicitly handling player movement
and turning instead, and they behave very close to how it is in Arena.
No jumping yet though, and sprinting is simply something for development
(I don't think it should stay in unless it's an option).