Moved "touched voxel" methods to relevant classes.
The Rect3D and Light classes now have their own methods for finding the
set of voxels that they touch, instead of having both versions stored in
the CLProgram. The Rect3D class was also moved from Math to Rendering
since that's closer to where it's being used. I'm not sure what I want
to do with the Sprite class yet.