Replaced solid color pop-ups with parchment.
Now that a couple pop-up textures can be generated, all of the pop-ups
that had been using placeholders have been switched to the intended
texture. Also moved around a bit of the ChooseNamePanel text and
increased the size of a space in font A by 1 pixel so it matches the
layout in the original game more closely.