Added brackets to unique_ptr array initializations.
The unique_ptr initializations that allocate a pointer to an array have
been secretly causing undefined behavior, but it only seems to be a
problem when the allocated objects are not primitives. This problem was
undetected because all the unique_ptr arrays I've been using so far have
been of type uint32_t.