Rough draft implementation of point lights.
A very rough draft, to be sure. I commented out the owner references to
simplify the lighting code and implemented basic ray casting for lights
in the kernel. It works, but the inefficiency of my code is starting to
show through, and the lights won't be finished until light owners are
implemented (so that sprites like street lamps can be fully lit by their
own light). More OpenCL refactoring will be on the horizon at some
point.