Removed dependency on OpenCL.
I'm moving away from OpenCL in favor of maintaining a software renderer
instead. My intention is to make it easier for me to work with the
rendering code and to encourage faster prototyping (since working with a
graphics API carries additional overhead with it, and I want to avoid
that this early in development). I've learned a lot using OpenCL,
though, and that's the important part.