Added default and decrypt modes to A.EXE unpacker.
Each "mode" in the executable unpacker determines how the bits in a bit
array are interpreted. The "default" mode simply selects between decrypt
or copy mode based on the current bit. "Decrypt" mode takes a compressed
byte and XORs a key against the lower nibble to put it into a
decompressed state.