More experimenting with executable unpacker.
I put the important part of decrypting into a lambda for readability,
and switched the loop construct to a more general one. The more I work
with this unpacker, the more it feels like I'm writing C89 code. It's a
little hard to keep track of everything because, at a glance, there
isn't a clear indication of how long it loops for, or exactly what
values are being incremented each loop... trademarks of a heavily
imperative solution.