Initial attempt at 3D software ray caster.
I was able to extend Lode Vandevenne's ray caster from 2D to 3D (again).
This time though, it's more robust (uses std::floor() properly, no more
dependence on epsilon), and it's implemented for the CPU instead of
OpenCL. Right now the texture colors are just placeholder UV
coordinates, and it's pretty slow overall, but that's to be expected.
I'm glad I found a way to use std::thread instead of having to add a
dependency on OpenMP.