Fix EOL check in text parsers
The default behavior of `getline` is to get the next line without the
EOL symbol(s). Hence, `line.at(0)` should in theory always throw an
exception for an empty line, since there would be no characters in the
string.

However, this doesn't happen on Windows. The EOL sequence on Windows
is '\r' '\n', but `getline` uses just '\n' as a delimiter, resulting
in '\r' being left in the string, hence making the code work.

The code doesn't work if the EOL sequence of the text files is just
LF. This fix should solve that, while still keeping the code
compatible with CRLF endings on Windows. Checking for '\n' is now
redundant.