Added compass functionality.
The compass animates now! On another note, the SLIDER.IMG texture itself
has a couple quirks, such as "N" being drawn in two places, and tick
marks erroneously being drawn where the second "NE" would be. It also
took me a while to figure out that I should use atan2(), not atan(), for
the angle of the player's direction.