Added VoxelGrid class, removed some others.
The VoxelGrid class should provide sufficient flexibility for
representing Arena's voxel IDs. It is a 3D array of IDs, and also has a
"voxel height" member because some of Arena's levels have "tall" voxels.
I removed a few of the existing voxel source files because they were
unused and were mostly for brainstorming.