Experimenting with tall voxel rendering.
Some of Arena's interior locations, like the Mage's Guild, have "tall"
voxels which are about 50% taller than a typical 1x1x1 cube. This
requires quite a few changes in the software renderer, and I've been
struggling a bit figuring out what all the new stepping values and
various lengths should be. Currently, it renders tall voxels for the
same floor the player is in, but it incorrectly renders cubes for all
other floors. Using std::fmod() also causes issues with some edge cases
for texture coordinates, and I'm not sure why that is.