Initial tests with 2.5D ray casting.
Although 3D ray casting would be more flexible, it's far too slow (even
with multi-threading), so I'm going to see how well 2.5D ray casting
turns out to be. This is the form of ray casting that draws "fake" 3D
worlds, and requires walls to be perpendicular to the ground. I would
assume that this is what Arena actually uses, to some extent.