Added perspective transformations to ray casting.
This adds much needed utility to the ray casting algorithm. I can
already get variable eye height just with this change. I think using
projections of 3D points as the basis for rendering columns will make it
a lot easier to add features. On another note, I am curious about
whether the wall rendering algorithm I figured out is actually identical
to what Arena uses, because it shares the same frequent "off-by-one"
artifacts on textures, which seem to be a result of involving various
rounding functions (ceil(), floor(), round()) with the projected
coordinates. It's definitely not from floating-point imprecision
(because I'm using doubles), so it is probably rounding-related.