Initial work on MIF format.
A MIF file describes map data. It appears to be used for main quest
dungeons, cities, building interiors, and probably the wilderness as
well. There are several kinds of data that may or may not be present in
each MIF file, so it will take some work to decode it properly.