First attempt at 2.5D flat rendering.
This is my attempt at applying the same idea behind wall rendering to
flat (i.e., sprite) rendering. It kind of works, in that the flats are
positioned properly and are generally textured correctly, but there are
several cases where it fails because of naive code, and the depth
testing with voxels isn't correct either. It's also lacking some typical
features like line clipping for flats that intersect the viewing plane,
so I'll need to investigate that. Anyway, it's off to a good start
overall, so that's nice. *Sigh*... I think all of these edge cases are
usually taken care of for us by APIs like OpenGL. I get to do it the
hard way!