Simplified some entity and camera code.
I originally designed most of the entity code from the top-down, but
after seeing that most of the abstraction I had planned is unnecessary,
I decided to revisit some of it from a bottom-up perspective. I think it
makes the code less general, which is good for now. Too much indirection
is not a good thing. Also renamed CharacterGender to Gender.