Minor change to fog calculations.
Moving the division into the std::min() makes it easier to see that the
upper bound on fog percent is 1.0. On another note, I wonder if the
texturing artifacts aren't from rounding errors, but because I'm not
using normalized device coordinates correctly. Some quick tests show
that the projected Z values are between 0.0 and 1.0 (which makes sense),
but in OpenGL, they should be between -1.0 and 1.0 (which makes less
sense, but I'll go with it).