Changed depth test to "less than or equal to".
It's important that two things sitting at the same Z distance get a
chance to be drawn, so that transparent sprites will be drawn correctly,
for example. The usual depth function in OpenGL is "GL_LEQUAL" as well.
Also added a depth check for wall drawing. Even though the ray casting
guarantees one wall intersection per screen column, this won't
necessarily be true with floor and ceiling rendering.