Added near/far planes, renamed viewDistance to fogDistance.
The near and far clipping planes will be used with the perspective
transformations for putting points on the screen properly (I don't think
the clipping is completely correct right now). The "viewDistance" member
was kind of ambiguous, so it's been renamed to "fogDistance", which
indicates how far the camera can see until the fog is completely thick.
Also commented out the 3D ray casting method since I won't be using it
now. I'll keep it around as a reference, though.