Moved test world creation to GameData class.
The majority of the test world is now created by the
GameData::createDefault() static method. I wanted to get this
functionality out of the ChooseAttributesPanel so it'd be more
organized. It's still not perfect, but it's better. Also changed some
std::unique_ptr members to values, since they don't need to be nullable.