Refined some callback arguments.
The Button class no longer has a default "Game" argument for click(),
and I removed the Game parameter from several buttons because they
simply don't need access to the whole universe. Some of them just change
a small part of the panel they're in. Having the arguments more visible
makes it easier to see what exactly the buttons are using.