Removed some unnecessary rendering methods.
I took out the set() methods and class members for some camera values in
the SoftwareRenderer because they can just go with the render() method
instead and decrease the amount of overall saved state. The parameters
for the render() method remind me a little bit more of OpenGL shaders
now. I think that's a reasonable model to follow in general. I like
passing arguments instead of saving state because stateless functions
are easier to wrap my mind around quickly.