Removed voxel height from VoxelGrid.
My previous understanding of Arena's voxel sizes was that in the Mages'
Guild, etc., the height of all voxels was equal to the height of voxels
on the main floor, but I now know this is false. Only voxels on the main
floor have a custom height for those particular locations, and all other
voxels (at the ground floor and above the main floor) are always 1x1x1
cubes. When rendering these custom locations, it should be enough to
simply specify a voxel data with a custom voxel height, and have the
relevant voxels point to that (unfortunately, it can't be generalized
for ALL voxels on the main floor, since ones on the perimeter of the
level can be normal sized instead of tall for some strange reason!).