Added far clipping to visible sprite determination.
The flats (i.e., sprites) are now thrown out of the visible flats list
if both of the vertical edges' Z values are greater than the far
plane's. This would mean the flat is completely obscured by fog. The
visible sprite determination algorithm and the sprite rendering both
need a lot of polishing, but I'll figure it out eventually.