Removed some entity abstraction.
The player doesn't really need to be an entity with an ID, so I made
them separate. All entities will be things in the world that are
non-player objects, and I think they will all have sprites. This makes
it a little easier to work with them instead of having to add conditions
for the player.