Added basic line clipping to flats.
This makes rendering the flats a bit more robust, as they no longer
disappear if any part of them is behind the near clipping plane.
However, the texture coordinates and depth calculation still become
skewed when the camera is very close to a flat, and I think I will need
to change some affine calculations to be perspective-correct instead.