Redesigned test world.
This new hand-placed city is more like what I had in mind for testing
with. Floor and ceiling IDs in voxel data objects have been separated
for more flexibility. The first voxel data object should now always be
the "empty voxel", which has 0 for all its IDs and is invisible when
rendering. Also switched the music to SUNNYDAY.XFM for a change from the
dreary overcast music. Also moved some of the world construction from
the ChooseAttributesPanel to the GameData::createDefault() method. Also
modified voxel data indices in the renderer to account for the first
voxel data object being the empty voxel.