Added basic automap functionality.
The automap now displays a top-down view of the map like in the original
game. It currently shows a yellow box for the player's position
(eventually an arrow instead). It also shows the location name at the
top. The compass buttons do nothing yet, but they will scroll the map
eventually.