Finished modern interface rough draft.
The player can now move and look around in modern interface mode with
WASD and the mouse (while holding the LMB or RMB). The rendered world
skews as the player looks up and down because it's using 2.5D ray
casting. Selecting things in the world isn't implemented yet for either
mode, so that will need to be done eventually.