Added sun direction method.
The voxels currently don't have any associated normal vectors, so the
sun direction doesn't do very much, but it will once I get partway
through the next iteration of the software renderer. The normals can be
trivially calculated based on which face of the voxel is being rendered,
and whether or not it's a back face.