Adjusted pixel sampling location.
Each pixel in the software renderer now retrieves its color from the
center of the pixel instead of the top left corner. This change isn't
really visible to the user, but on another note, it seems to have fixed
the issue with noisy artifacts along the edges of floors (due to edges
sharing Z values with adjacent voxels), and I'm not sure why.