Experimenting with collision detection.
I have a simple working model of collision detection in the XZ plane
now. It only looks at whether a voxel is air or not, so half-voxels like
the bridge are currently treated as a solid wall. Later, the player's
collision should be done with an axis-aligned bounding box instead of a
point.