Changed program icon.
We're now using a new icon! It seems decent enough for now. and it can
always be changed in the future. Also added a little bit of code so PPM
icons will use black as the transparent color (since having PNG support
just for one icon seems a bit unnecessary to me).