Added text box shadow functionality.
This has been a desired feature for several months. Text boxes now have
a shadow texture automatically generated during construction that is
simply a copy of the main texture but with another text color. I figured
this was easier than resizing the main texture based on the shadow's XY
offset.