Added CITYDATA.00 parser.
CITYDATA.00 contains the names and screen coordinates of all locations
for every province on the world map. There is some unknown data, and I'm
not sure if it could be the random seeds for each location, or maybe
coordinates for random dungeons, or maybe something else. Anyway, there
should be enough data now to draw the icons and names of each location
on the world map.