Replaced some hardcoded filenames with A.EXE segments.
This seemed like a reasonable idea because the province map filenames
appear in the same order in A.EXE as their province names do, so their
indices are reusable as a result. Also added new String methods for
converting an ASCII string to uppercase or lowercase, since the texture
manager expects filenames to be uppercase, and most of the filenames in
A.EXE are lowercase.